refactor: ImportZipsの参照先をscripts配下に統一しapplication側実体を削除する
レビュー指摘の通り src/application/app/importZips/ImportZips.js を削除し、
旧パス参照を残すと再発しやすい構造だったためテストコードの参照先を
scripts/ImportZips.js に統一した。
これにより配置方針をコード上で明確化し、将来の保守時に参照ゆれが発生しないようにするため。
