--- src/application/app/importZips/ImportZips.js
@@ -1,1 +1,0 @@
-module.exports = require('../../../../scripts/ImportZips');
